Move websocket logic to separate file
